nfs: split record parsers into different files
